+++ src/main/java/com/skycat/mystical/command/MysticalCommandHandler.java
@@ -54,6 +54,9 @@ public class MysticalCommandHandler implements CommandRegistrationCallback {
                 .requires(Permissions.require("mystical.command.mystical", true))
                 // TODO: send some info
                 .build();
+        var credits = literal("credits")
+                .executes(this::creditsCommand)
+                .build();
         var spell = literal("spell")
                 .requires(Permissions.require("mystical.command.mystical.spell", true))
                 .build();
@@ -172,6 +175,11 @@ public class MysticalCommandHandler implements CommandRegistrationCallback {
         */
 
 
+    }
+
+    private int creditsCommand(CommandContext<ServerCommandSource> context) {
+        context.getSource().sendFeedback(Utils.translatableSupplier("text.mystical.command.mystical.credits"), false);
+        return Command.SINGLE_SUCCESS;
     }
 
     private int deleteSpellAllCommand(CommandContext<ServerCommandSource> context) throws CommandSyntaxException {